Refactor config.py into config folder
