+++ core/db.py
@@ -1,6 +1,6 @@
 import sqlite3
 from core.log import log
-from config import SQLITE3_DB_NAME, DB_TABLE_REMINDERS_NAME
+from config.config import SQLITE3_DB_NAME, DB_TABLE_REMINDERS_NAME
 
 
 # Database definitions
